refactor: sorting parameters to maintain a unified API style
