+++ application/src/main/java/run/halo/app/extension/Comparators.java
@@ -17,4 +17,15 @@ public enum Comparators {
         return asc ? comparator : comparator.reversed();
     }
 
+    /**
+     * Get a nulls comparator.
+     *
+     * @param isAscending is ascending
+     * @return if ascending, return nulls low, else return nulls high
+     */
+    public static Comparator<Object> nullsComparator(boolean isAscending) {
+        return isAscending
+            ? org.springframework.util.comparator.Comparators.nullsLow()
+            : org.springframework.util.comparator.Comparators.nullsHigh();
+    }
 }
